refactor: move fluctuations.
