+++ ptahlmud/datastack/__init__.py
@@ -5,7 +5,6 @@ This package provides tools to fetch, store and serve market data.
 
 from ptahlmud.datastack.clients.binance_client import BinanceClient
 from ptahlmud.datastack.clients.remote_client import RemoteClient
-from ptahlmud.datastack.fluctuations import Fluctuations
 from ptahlmud.datastack.fluctuations_service import FluctuationsService, FluctuationsSpecs
 
-__all__ = ["BinanceClient", "Fluctuations", "FluctuationsService", "FluctuationsSpecs", "RemoteClient"]
+__all__ = ["BinanceClient", "FluctuationsService", "FluctuationsSpecs", "RemoteClient"]
