+++ ptahlmud/datastack/testing/fluctuations.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import numpy as np
 import pandas as pd
 
-from ptahlmud.datastack.fluctuations import Fluctuations
+from ptahlmud.core.fluctuations import Fluctuations
 from ptahlmud.types import Period
