+++ tests/datastack/test_fluctuations_service.py
@@ -2,8 +2,8 @@ from datetime import datetime, timedelta
 
 import pytest
 
+from ptahlmud.core.fluctuations import Fluctuations
 from ptahlmud.datastack.clients.remote_client import RemoteClient
-from ptahlmud.datastack.fluctuations import Fluctuations
 from ptahlmud.datastack.fluctuations_service import CustomOperation, FluctuationsService, FluctuationsSpecs
 from ptahlmud.datastack.testing.fluctuations import generate_fluctuations
 from ptahlmud.types import Period
